requirements.txt
Flask
selenium
Pillow
requests


